simplify the original framework
