+++ utils/__init__.py
@@ -1,6 +1,4 @@
 from .accumulator import Accumulator
-from .ctrl_panel import ControlPanel
-from .experiment import Experiment
 from .history import History
 from .process import Process
 from .thread import Thread
@@ -9,7 +7,6 @@ from .func import log_tools as ltools
 from .func import pytools as ptools
 from .func import tensor_tools as tstools
 from .func import torch_tools as ttools
-from .new_ctrl_panel import ControlPanel as NewControlPanel
-from .new2_ctrl_panel import New2ControlPanel
-from .new2_experiment import New2Experiment
+from .ctrl_panel import New2ControlPanel
+from .experiment import New2Experiment
