--- utils/new_ctrl_panel.py
@@ -1,166 +1,0 @@
-import json
-import os.path
-
-import jsonref
-import pandas as pd
-import torch
-from jsonref import JsonRef
-
-from config.init_cfg import init_log, init_train_settings, init_hps
-from .new_experiment import Experiment
-from .func import pytools as ptools
-
-
-def resolve_jsonref(obj):
-    """用于处理JsonRef对象"""
-    if isinstance(obj, JsonRef):
-        return resolve_jsonref(obj.__subject__)  # 解引用
-    elif isinstance(obj, dict):
-        return {k: resolve_jsonref(v) for k, v in obj.items()}
-    elif isinstance(obj, list):
-        return [resolve_jsonref(item) for item in obj]
-    else:
-        return obj
-
-
-
-class ControlPanel:
-    """控制台类负责读取、管理动态运行参数、超参数组合，以及实验对象的提供"""
-
-    def __init__(self, datasource, net_type, is_train, cfg_root=os.path.join(".", "config")):
-        """控制面板
-
-        负责读取、管理动态运行参数、超参数组合，生成超参数配置文件路径、日志文件存储路径以及网络文件存储目录，提供实验Experiment对象。
-        迭代每次提供一个实验对象，包含有单次训练的超参数组合以及动态运行参数组合，训练过程中不可改变，每次迭代后会更新运行配置参数。
-        :param datasource: 训练数据来源
-        :param net_type: 训练所用模型类
-        :param cfg_root: 运行配置文件路径
-        """
-        net_name = net_type.__name__.lower()
-        self.net_type = net_type
-        # 生成运行动态配置
-        self.__rcp = os.path.join(cfg_root, net_name, f'settings.json')  # 运行配置json文件路径
-        # 读取运行配置
-        ptools.check_path(self.__rcp, init_train_settings)
-        self.__read_runtime_cfg()
-        # 生成其他路径
-        self.__hcp = os.path.join(cfg_root, net_name, f'hyper_param_s.json')  # 网络训练超参数文件路径
-        log_root = self.cfg_dict['log_root']
-        self.__mlp = os.path.join(log_root, net_name, f'metric_log.csv')  # 指标日志文件存储路径
-        self.__plp = os.path.join(log_root, net_name, f'perf_log.csv')  # 性能日志文件存储路径
-        self.__np = os.path.join(log_root, net_name, f'trained_net/')  # 训练成果网络存储路径
-        self.__pp = os.path.join(log_root, net_name, f'imgs/')  # 历史趋势图存储路径
-        # 路径检查
-        ptools.check_path(self.__hcp, init_hps)
-        ptools.check_path(self.__mlp, init_log)
-        ptools.check_path(self.__plp, init_log)
-        ptools.check_path(self.__np)
-        ptools.check_path(self.__pp)
-        # 设置随机种子
-        torch.random.manual_seed(self['random_seed'])
-        # # 读取实验编号
-        self.__read_expno()
-        self.dao_ds = datasource
-        self.is_train = is_train
-
-    def __read_expno(self):
-        """读取实验编号
-        从日志中读取最后一组实验数据的实验编号，从而推算出即将进行的所有实验组对应的编号。
-        本函数将会创建self.exp_no以及self.last_expno属性。
-        :return: None
-        """
-        # 读取性能日志的最后一项，取出实验编号作为本次实验编号的参考
-        try:
-            log = pd.read_csv(self.__plp)
-            exp_no = log.iloc[-1]['exp_no'] + 1
-        except Exception as _:
-            exp_no = 1
-        assert exp_no > 0, f'训练序号需为正整数，但读取到的序号为{exp_no}'
-        self.exp_no = int(exp_no)
-        # 计算总共需要进行的实验组数
-        with open(self.__hcp, 'r', encoding='utf-8') as cfg:
-            hyper_params = json.load(cfg)
-            n_exp = 1
-            for v in hyper_params.values():
-                n_exp *= len(v)
-        # 最后一组实验的实验编号
-        self.last_expno = self.exp_no + n_exp - 1
-    #
-    # def build_dao_ds(self, hyper_params):
-    #     pos_only_args, pos_or_kwargs, kwargs_only, _ = ptools.get_signature(self.dao_ds)
-    #     args = [hyper_params.pop(poa) for poa in pos_only_args + pos_or_kwargs]
-    #     kwargs = {ko: hyper_params.pop(ko) for ko in kwargs_only}
-    #     ds_kwargs = self.cfg_dict.pop("ds_kwargs")
-    #     return self.dao_ds(*args, **kwargs, module=self.net_type, config=ds_kwargs, is_train=self.is_train)
-
-    def __iter__(self):
-        """迭代
-        每次提供一个实验对象，包含有单次训练的超参数组合以及动态运行参数组合，训练过程中不可改变。
-        每次迭代会更新运行配置参数。
-        :return: None
-        """
-        with open(self.__hcp, 'r', encoding='utf-8') as cfg:
-            hyper_params = json.load(cfg)
-            hp_keys = hyper_params.keys()
-        for hps in ptools.permutation([], *hyper_params.values()):
-            hyper_params = {k: v for k, v in zip(hp_keys, hps)}
-            # nb_kwargs = self.cfg_dict.pop("nb_kwargs")
-            # t_kwargs = self.cfg_dict.pop("t_kwargs")
-            # 构造实验对象
-            print(
-                f'\r---------------------------'
-                f'实验{self.exp_no}号/{self.last_expno}号'
-                f'---------------------------'
-            )
-            cur_exp = Experiment(
-                self.exp_no, self.dao_ds, self.net_type,
-                hyper_params, self.cfg_dict, self.is_train,
-                self.__pp, self.__mlp, self.__plp, self.__np
-            )
-            # data = self.__datasource(
-            #     hps['d_size'], hps['s_quantity'], hps['s_size'], hps["test_which"],
-            #     hps["flip_prob"], hps["rotate_prob"], hps["add_noise_prob"],
-            #     train_which=hps['train_which'], module=self.net_type, config=ds_kwargs, is_train=True
-            # )
-            # data = self.build_dao_ds(hyper_params)
-            # net_builder = NetBuilder(self.net_type, nb_kwargs)
-            # trainer = Trainer(net_builder, data.get_criterion_a(), t_kwargs, hyper_params)
-            yield cur_exp
-            # 读取运行动态参数
-            self.__read_runtime_cfg()
-            if torch.cuda.is_available():
-                torch.cuda.empty_cache()
-            # 更新实验编号
-            self.exp_no += 1
-
-    def __getitem__(self, item):
-        """获取控制面板中的运行配置参数。
-
-        :param item: 运行配置参数名称
-        :return: 运行配置参数值
-        """
-        assert item in self.cfg_dict.keys(), f'设置文件中不存在{item}参数！'
-        return self.cfg_dict[item]
-
-    def __read_runtime_cfg(self):
-        """读取运行配置并更新或者赋值"""
-        with open(self.__rcp, 'r', encoding='utf-8') as config:
-            config_dict = jsonref.load(config, merge_props=True)
-            config_dict = resolve_jsonref(config_dict)
-            self.cfg_dict = config_dict
-            # if hasattr(self, 'cfg_dict'):
-            #     # 更新运行动态参数
-            #     # assert config_dict.keys() == self.cfg_dict.keys(), '在运行期间，不允许添加新的运行设置参数！'
-            #     for k, v in config_dict.items():
-            #         self.cfg_dict[k] = v
-            # else:
-            #     # 首次获取运行动态参数
-            #     self.cfg_dict = config_dict
-
-    @property
-    def runtime_cfg(self):
-        return self.cfg_dict
-
-    @property
-    def device(self):
-        return torch.device(self['device'])
